requirements.txt
requests
moviepy
pytube3
youtube_transcript_api
SpeechRecognition
pydub
deep-translator
transformers
huggingsound 
librosa
nltk
evaluate
rouge
spacy
matplotlib
seaborn
scikit-learn
torch
validators
Flask
Flask-Cors
numpy
regex
tqdm
urllib3
python-dotenv
os
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl